+++ packages/core/src/hooks/useTokenAllowance.ts
@@ -1,26 +1,29 @@
 import { BigNumber } from '@ethersproject/bignumber'
+import { Contract } from 'ethers'
 import { ERC20Interface } from '../constants'
 import { QueryParams } from '../constants/type/QueryParams'
 import { Falsy } from '../model/types'
-import { useContractCall } from './useContractCall'
+import { useCall } from './useCall'
 
+/**
+ * @public
+ */
 export function useTokenAllowance(
   tokenAddress: string | Falsy,
   ownerAddress: string | Falsy,
   spenderAddress: string | Falsy,
   queryParams: QueryParams = {}
 ): BigNumber | undefined {
-  const [allowance] =
-    useContractCall(
+  const { value: allowance } =
+    useCall(
       ownerAddress &&
         spenderAddress &&
         tokenAddress && {
-          abi: ERC20Interface,
-          address: tokenAddress,
+          contract: new Contract(tokenAddress, ERC20Interface),
           method: 'allowance',
           args: [ownerAddress, spenderAddress],
         },
       queryParams
-    ) ?? []
-  return allowance
+    ) ?? {}
+  return allowance?.[0]
 }