+++ packages/core/src/providers/chainState/common/callsReducer.ts
@@ -2,6 +2,9 @@ import { addressEqual, ChainId } from '../../..'
 
 export type Action = AddCall | RemoveCall
 
+/**
+ * @public
+ */
 export interface RawCall {
   chainId: ChainId
   address: string