+++ packages/core/src/providers/chainState/common/chainStateReducer.ts
@@ -1,5 +1,8 @@
 import { ChainState } from './model'
 
+/**
+ * @public
+ */
 export interface State {
   [chainId: number]:
     | {