+++ packages/docs/docusaurus.config.js
@@ -16,6 +16,10 @@ const config = {
   organizationName: 'TrueFiEng', // Usually your GitHub org/user name.
   projectName: 'useDApp', // Usually your repo name.
 
+  plugins: [
+    './plugins/webpack-plugin.js'
+  ],
+
   presets: [
     [
       'classic',
@@ -24,7 +28,7 @@ const config = {
         docs: {
           sidebarPath: require.resolve('./sidebars.js'),
           // Please change this to your repo.
-          editUrl: 'https://github.com/TrueFiEng/useDApp/tree/docosaurus/packages/docs/',
+          editUrl: 'https://github.com/TrueFiEng/useDApp/tree/master/packages/docs/',
           remarkPlugins: [
             [require('@docusaurus/remark-plugin-npm2yarn'), {sync: true}],
           ],
@@ -33,7 +37,7 @@ const config = {
           showReadingTime: true,
           // Please change this to your repo.
           editUrl:
-            'https://github.com/TrueFiEng/useDApp/tree/docosaurus/packages/docs/',
+            'https://github.com/TrueFiEng/useDApp/tree/master/packages/docs/',
         },
         theme: {
           customCss: require.resolve('./src/css/custom.css'),
@@ -91,6 +95,10 @@ const config = {
               {
                 label: 'GitHub',
                 href: 'https://github.com/TrueFiEng/useDApp/',
+              },
+              {
+                label: 'Discord',
+                href: 'https://discord.gg/cSSmtdq7jr',
               },
             ],
           },