+++ packages/docs/src/examples/ConnectingToNetwork.tsx
@@ -1,0 +1,36 @@
+import React from 'react'
+import ReactDOM from 'react-dom'
+import { formatEther } from '@ethersproject/units'
+import { Mainnet, DAppProvider, useEtherBalance, useEthers, Config } from '@usedapp/core'
+
+const config: Config = {
+    readOnlyChainId: Mainnet.chainId,
+    readOnlyUrls: {
+      [Mainnet.chainId]: 'https://mainnet.infura.io/v3/57fc2c19095745e59ab96a4aa87dada8',
+    },
+}
+  
+ReactDOM.render(
+    <DAppProvider config={config}>
+      <App />
+    </DAppProvider>,
+    document.getElementById('root')
+)
+
+const STAKING_CONTRACT = '0x00000000219ab540356cBB839Cbe05303d7705Fa'
+
+export function App() {
+  const { activateBrowserWallet, deactivate, account } = useEthers()
+  const userBalance = useEtherBalance(account)
+  const stakingBalance = useEtherBalance(STAKING_CONTRACT)
+  
+  return (
+    <div>
+      {!account && <button onClick={activateBrowserWallet}> Connect </button>}
+      {account && <button onClick={deactivate}> Disconnect </button>}
+      {account && <p>Account: {account}</p>}
+      {userBalance && <p>Ether balance: {formatEther(userBalance)} ETH </p>}
+      {stakingBalance && <p>ETH2 staking balance: {formatEther(stakingBalance)} ETH </p>}
+    </div>
+  )
+}