+++ packages/example/src/pages/WalletConnect.tsx
@@ -1,5 +1,6 @@
-import React from 'react'
+import React, { useState } from 'react'
 import { formatEther } from '@ethersproject/units'
+import { Web3Provider } from '@ethersproject/providers'
 import { useEtherBalance, useEthers } from '@usedapp/core'
 import { Container, ContentBlock, ContentRow, MainContent, Section, SectionRow } from '../components/base/base'
 import { Label } from '../typography/Label'
@@ -11,14 +12,36 @@ import WalletConnectProvider from '@walletconnect/web3-provider'
 const STAKING_CONTRACT = '0x00000000219ab540356cBB839Cbe05303d7705Fa'
 
 export function WalletConnect() {
-  const { account, activate } = useEthers()
+  const { account, activate, chainId, deactivate, library } = useEthers()
+  const [signedMessage, setSignedMessage] = useState('')
 
   async function onConnect() {
-    const provider = new WalletConnectProvider({
-      infuraId: '57fc2c19095745e59ab96a4aa87dada8',
-    })
-    await provider.enable()
-    activate(provider)
+    try {
+      const provider = new WalletConnectProvider({
+        infuraId: 'd8df2cb7844e4a54ab0a782f608749dd',
+      })
+      await provider.enable()
+      activate(provider)
+    } catch (error) {
+      console.error(error)
+    }
+  }
+
+  async function onDisconnect() {
+    deactivate()
+    localStorage.removeItem('walletconnect')
+    setSignedMessage('')
+  }
+
+  async function onSign() {
+    const msg = 'I sign Wallet Connect test message on @usedapp'
+    const provider = library as Web3Provider
+    try {
+      const signedMsg = await provider.getSigner().signMessage(msg)
+      setSignedMessage(signedMsg)
+    } catch (error) {
+      console.error(error)
+    }
   }
 
   const userBalance = useEtherBalance(account)
@@ -30,13 +53,12 @@ export function WalletConnect() {
         <Section>
           <SectionRow>
             <Title>WalletConnect Usage Example</Title>
-            <Button onClick={onConnect}>Connect</Button>
+            <Button onClick={account ? onDisconnect : onConnect}>{account ? 'Disconnect' : 'Connect'}</Button>
           </SectionRow>
           <ContentBlock>
-            {stakingBalance && (
+            {chainId && account && (
               <ContentRow>
-                <Label>ETH2 staking contract holds:</Label> <TextInline>{formatEther(stakingBalance)}</TextInline>{' '}
-                <Label>ETH</Label>
+                <Label>Active Chain ID:</Label> <TextInline>{chainId}</TextInline>{' '}
               </ContentRow>
             )}
             {account && (
@@ -49,7 +71,24 @@ export function WalletConnect() {
                 <Label>Ether balance:</Label> <TextInline>{formatEther(userBalance)}</TextInline> <Label>ETH</Label>
               </ContentRow>
             )}
+            {stakingBalance && (
+              <ContentRow>
+                <Label>ETH2 staking contract holds:</Label> <TextInline>{formatEther(stakingBalance)}</TextInline>{' '}
+                <Label>ETH</Label>
+              </ContentRow>
+            )}
+            {signedMessage && account && (
+              <ContentRow>
+                <Label>Signed message signature:</Label>{' '}
+                <TextInline style={{ overflowWrap: 'break-word' }}>{signedMessage}</TextInline>{' '}
+              </ContentRow>
+            )}
           </ContentBlock>
+          {account && (
+            <SectionRow style={{ marginTop: '20px', display: 'flex', justifyContent: 'flex-end' }}>
+              <Button onClick={onSign}>Sign message</Button>
+            </SectionRow>
+          )}
         </Section>
       </Container>
     </MainContent>